requirements.txt
mysql-connector
mysqlclient
pyqt5
requests
pyinstaller